summary

+++ TURV/models.py
@@ -133,7 +133,7 @@ class TabelItem(models.Model):
     hours31 = models.CharField(max_length=4, verbose_name='Часы31', null = True, blank=True)
 
 #Итоги видов времени
-    sHours1 = models.IntegerField(verbose_name='Явки (Я)', help_text='Явки', null = True, blank=True)
+    sHours1 = models.FloatField(verbose_name='Явки (Я)', help_text='Явки', null = True, blank=True)
     sHours2 = models.IntegerField(verbose_name='Ночные (Н)', null = True, blank=True)
     sHours3 = models.IntegerField(verbose_name='Работа в выходные и празд. (РВ)', null = True, blank=True)
     sHours4 = models.FloatField(verbose_name='Сверхурочные (С)', null = True, blank=True)
